test-requirements.txt
nose>=1.3,<2.0
spur.local>=0.3,<0.4
tempman>=0.1.3,<0.2
slimit>=0.8.1,<0.9
requests>=2.4.3,<3.0
